updating with more feedback
