+++ src/Microsoft.Azure.SignalR.Common/ServiceConnections/ClientConnectionScopeInternal.cs
@@ -1,28 +1,12 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using Microsoft.Azure.SignalR.Common.Utilities;
 using System;
 using System.Diagnostics;
-using System.Dynamic;
-using System.Threading;
 
 namespace Microsoft.Azure.SignalR
 {
-
-    internal class ScopePropertiesAccessor<TProps>
-    {
-        // Use async local with indirect reference to TProps to allow for deep cleanup
-        private static readonly AsyncLocal<ScopePropertiesAccessor<TProps>> s_currentAccessor = new AsyncLocal<ScopePropertiesAccessor<TProps>>();
-
-        internal protected static ScopePropertiesAccessor<TProps> Current
-        {
-            get => s_currentAccessor.Value;
-            set => s_currentAccessor.Value = value;
-        }
-
-        internal TProps Properties { get; set; }
-    }
-
     /// <summary>
     /// Represents a disposable scope able to carry connection properties along with the execution context flow
     /// </summary>
